Fix DiffUtillCallback Support - now relies on a DiffItem interface
