+++ library/src/main/java/com/vicidroid/amalia/ui/recyclerview/diff/DiffItem.kt
@@ -1,0 +1,9 @@
+package com.vicidroid.amalia.ui.recyclerview.diff
+
+/**
+ * Used in conjunction with [RecyclerItemDiffCallback.areItemsTheSame]
+ * Leverage this interface for your data
+ */
+interface DiffItem {
+    val diffId: Long
+}
